Made different files for different processes
