+++ src/main.rs
@@ -1,33 +1,28 @@
-use self::models::Blockchain;
-use diesel::prelude::*;
-use dotenvy::dotenv;
-use jsonrpsee::ws_client::{WsClient, WsClientBuilder};
-use models::NewBlockchain;
-use rocket::{self, launch, routes};
-use std::env;
+use jsonrpsee::ws_client::WsClient;
 use std::io::{self, Write};
-use std::thread;
 
 mod blockchain_api;
 mod blockchain_data;
 pub use blockchain_data::*;
+pub use blockchain_api::*;
+pub use connection::*;
 
+pub mod connection;
 pub mod models;
 pub mod schema;
 
-pub fn establish_connection() -> PgConnection {
-    dotenv().ok();
-    let database_url = env::var("DATABASE_URL").expect("DATABASE_URL must be set");
-    PgConnection::establish(&database_url)
-        .unwrap_or_else(|_| panic!("Error connecting to {}", database_url))
-}
+#[tokio::main]
+async fn main() {
+    println!("🚀 Hello! Welcome to the Blockchain World 🌐\n");
+    println!("📋 Choose an option:\n1️⃣  Start the Rocket Server\n2️⃣  Store Blockchain Data\n");
 
-/// Establish WebSocket connection and handle errors
-async fn establish_ws_connection(endpoint: &str) -> Result<WsClient, String> {
-    WsClientBuilder::default()
-        .build(endpoint)
-        .await
-        .map_err(|_| "Failed to connect. Please enter a valid WebSocket endpoint.".to_string())
+    let user_input = user_input().trim().parse::<u32>().unwrap_or(0);
+
+    match user_input {
+        1 => rocket_launch().await,
+        2 => store_blockchain().await,
+        _ => println!("❌ Invalid option. Please restart and choose 1 or 2."),
+    }
 }
 
 /// Handle user input for the WebSocket endpoint
@@ -41,20 +36,6 @@ fn get_websocket_endpoint() -> String {
         .expect("Failed to read line");
 
     endpoint.trim().to_string()
-}
-
-// Store the name in the database
-fn store_blockchain(blockchain: &str) {
-    let new_blockchain = NewBlockchain {
-        blockchain_name: blockchain,
-    };
-
-    diesel::insert_into(schema::blockchain_info::table)
-        .values(&new_blockchain)
-        .get_result::<Blockchain>(&mut establish_connection())
-        .expect("💥 Error saving new blockchain");
-
-    println!("✅ Blockchain name successfully stored in the database! 🗄️");
 }
 
 /// Handle user input for the selected option
@@ -83,42 +64,19 @@ async fn fetch_blockchain_name(client: WsClient) {
     println!("Do you want to store in the database? ");
 
     let user_input = user_input();
-
     let command: String = user_input.to_lowercase().trim().parse().unwrap();
-
     let blockchain = get_blockchain_name.unwrap();
 
     match command.as_ref() {
-        "store" => store_blockchain(&blockchain),
+        "store" => store_db(&blockchain),
         "exit" => println!("👋 Goodbye!"),
         _ => println!("❗ Not a recognized keyword. Try again!"),
     }
 }
 
-// Take the input from the user
-fn user_input() -> String {
-    let mut command = String::new();
-    io::stdin().read_line(&mut command).unwrap();
-    command
-}
-
-// Rocket server launch configuration
-async fn rocket() {
-    let _ = rocket::build()
-        .mount("/", blockchain_api::rocket_routes())
-        .launch()
-        .await;
-}
-
-#[tokio::main]
-async fn main() {
-    println!("🚀 Hello! Welcome to the Blockchain World 🌐");
-
-    // Start the Rocket server in a separate thread
-    tokio::spawn(async {
-        rocket().await;
-    });
-
+// Check the Data and store in the Blockchain
+async fn store_blockchain(){
+    println!("💾 Preparing to store blockchain data...");
     let endpoint = get_websocket_endpoint();
 
     // Check if endpoint starts with "ws"
@@ -142,9 +100,11 @@ async fn main() {
     } else {
         println!("⚠️ Invalid WebSocket endpoint. It should start with 'ws://' or 'wss://'.");
     }
+}
 
-    // Simple way to keep the main thread alive
-    loop {
-        tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-    }
+// Take the input from the user
+fn user_input() -> String {
+    let mut command = String::new();
+    io::stdin().read_line(&mut command).unwrap();
+    command
 }
